Binomial expansion

+++ IrrationalNumbers/IrrationalNumbers.Logic/ParameterNormalizationResult.cs
@@ -4,6 +4,6 @@
     {
         public BigDecimal NormalizedParameter { get; set; }
 
-        public double FinalMultiplier { get; set; } = 1;
+        public BigDecimal FinalMultiplier { get; set; } = 1;
     }
 }
